+++ CHANGELOG.md
@@ -4,6 +4,11 @@ CHANGELOG
 master
 ------
 
+* todo...
+
+v1.2.0
+------
+
 * Improve message in case of encryption error
 * Add command to encrypt secrets
 * Drop support for PHP 7.1